requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.10.1
bleach==6.0.0
blessed==1.20.0
blinker==1.6.3
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.15
cleo==2.0.1
click==8.1.7
codeboxapi==0.0.19
codeinterpreterapi==0.0.14
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
crashtest==0.4.1
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
dulwich==0.21.6
emoji==2.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
fastapi==0.103.2
fastjsonschema==2.18.0
ffmpy==0.3.1
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
ghp-import==2.1.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.4
googleapis-common-protos==1.61.0
gradio==4.0.2
gradio_client==0.7.0
greenlet==3.0.0
grpcio==1.59.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
identify==2.5.31
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.3
install==1.3.5
installer==0.7.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==24.2.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.323
langchain-core==0.0.6
langchainhub==0.1.13
langdetect==1.0.9
langsmith==0.0.66
litellm==0.8.6
llama-index==0.8.49
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.1
mkdocs==1.5.3
mkdocs-material==9.4.6
mkdocs-material-extensions==1.3
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.4
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.16.0
ooba==0.0.21
open-interpreter==0.1.10
openai==0.28.1
openpyxl==3.1.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
orjson==3.9.10
oscrypto==1.3.0
overrides==7.4.0
packaging==23.2
paginate==0.5.6
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkginfo==1.9.6
platformdirs==3.11.0
poetry-core==1.7.0
poetry-plugin-export==1.5.0
posthog==3.0.2
pre-commit==3.5.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.3.1
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==3.1.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-editor==1.0.4
python-iso639==2023.6.15
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rapidfuzz==2.15.2
readchar==4.0.5
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.3.0
snowflake-sqlalchemy==1.5.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.49
stack-data==0.6.2
starlette==0.27.0
streamlit==1.27.2
streamlit-authenticator==0.2.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.14.1
tokentrim==0.1.13
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
trove-classifiers==2023.9.19
typer==0.9.0
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
unstructured==0.10.19
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
virtualenv==20.24.5
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
wget==3.2
widgetsnbextension==4.0.9
wikipedia==1.4.0
wrapt==1.15.0
xattr==0.10.1
xmltodict==0.13.0
yarl==1.9.2
yaspin==3.0.1
zipp==3.17.0
